+++ History.md
@@ -2,6 +2,7 @@
 ==================
   * Support 'include-unloaded-js' in proxy mode (https://github.com/tntim96/JSCover/issues/93)
   * Fix inclusion logic for unloaded-js (https://github.com/tntim96/JSCover/issues/93)
+  * Improve error message when invalid CL option supplied (https://github.com/tntim96/JSCover/issues/96)
 
 1.0.3 / 2013-08-25
 ==================